scripts/requirements.txt
neo4j>=5.19.0
python-dotenv>=1.0.0
langchain-openai>=0.1.0
pinecone>=5.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
python-dateutil>=2.9.0.post0
requests>=2.32.0

